Enable the test module to run
note this does nothing really.  The original as far back as 2015 was turned off.  It only stubbed looking at files.  The original java files were long since deleted.  I'd suspect one would want to see these actually format and needs a lot more.  However, the module is now enabled so at very least the rails work for potentially running tests.  In this currently there are 2 types of projects.  The one is a multi module project, it doesn't work so that one is disabled.  The reason is that its not a java project at parent and the plugin doesn't work that way anyways, its per module but left it anyways.